refactor: migrate monetary fields from f64 to rust_decimal::Decimal
- Domain: Pricing, UsageRecord, ApiKey spend_limit → Decimal
- DB layer: wallet/recharge/billing monetary reads/writes → Decimal
- Admin handlers: all billing/wallet response types → Decimal
- Cache: compute_gate_status, gate/balance methods → Decimal
- Config loader: seed pricing types → Decimal
- All Decimal fields use #[serde(with = "rust_decimal::serde::float")]

+++ src/admin/me.rs
@@ -4,6 +4,7 @@ use axum::extract::{Path, State};
 use axum::http::HeaderMap;
 use axum::Json;
 use chrono_tz::Tz;
+use rust_decimal::Decimal;
 use serde::Deserialize;
 use serde_json::Value;
 
@@ -182,8 +183,8 @@ pub(crate) struct CreateMyKeyReq {
     name: Option<String>,
     enabled: Option<bool>,
     expires_at: Option<String>,
-    #[serde(default)]
-    spend_limit: Option<f64>,
+    #[serde(default, with = "rust_decimal::serde::float_option")]
+    spend_limit: Option<Decimal>,
     #[serde(default)]
     allowed_models: Option<Vec<String>>,
 }
@@ -222,8 +223,8 @@ pub(crate) struct UpdateMyKeyReq {
     name: Option<String>,
     enabled: Option<bool>,
     expires_at: Option<String>,
-    #[serde(default)]
-    spend_limit: Option<f64>,
+    #[serde(default, with = "rust_decimal::serde::float_option")]
+    spend_limit: Option<Decimal>,
     #[serde(default)]
     allowed_models: Option<Vec<String>>,
 }
